Module signature::hazmat

Hazardous Materials: low-level APIs which can be insecure if misused.

The traits in this module are not generally recommended, and should only be used in special cases where they are specifically needed.

Using them incorrectly can introduce security vulnerabilities. Please carefully read the documentation before attempting to use them.

To use them, enable the hazmat-preview crate feature. Note that this feature is semi-unstable and not subject to regular 1.x SemVer guarantees. However, any breaking changes will be accompanied with a minor version bump.

Traits

PrehashSigner

Sign the provided message prehash, returning a digital signature.

PrehashVerifier

Verify the provided message prehash using Self (e.g. a public key)

RandomizedPrehashSigner

Sign the provided message prehash using the provided external randomness source, returning a digital signature.