signature::hazmat

Trait PrehashVerifier

pub trait PrehashVerifier<S> {
    // Required method
    fn verify_prehash(&self, prehash: &[u8], signature: &S) -> Result<(), Error>;
}

Verify the provided message prehash using Self (e.g. a public key)

Required Methods

fn verify_prehash(&self, prehash: &[u8], signature: &S) -> Result<(), Error>

Use Self to verify that the provided signature for a given message prehash is authentic.

The prehash parameter should be the output of a secure cryptographic hash function.

Returns Error if it is inauthentic or some other error occurred, or otherwise returns Ok(()).

⚠️ Security Warning

If prehash is something other than the output of a cryptographically secure hash function, an attacker can potentially forge signatures by solving a system of linear equations.