signature

Module signer

Traits for generating digital signatures.

Traits

AsyncDigestSigner

Asynchronously sign the given prehashed message Digest using Self.

AsyncRandomizedSigner

Sign the given message using the provided external randomness source.

AsyncSigner

Asynchronously sign the provided message bytestring using Self (e.g. client for a Cloud KMS or HSM), returning a digital signature.

DigestSigner

Sign the given prehashed message Digest using Self.

MultipartSigner

Equivalent of Signer but the message is provided in non-contiguous byte slices.

RandomizedDigestSigner

Combination of DigestSigner and RandomizedSigner with support for computing a signature over a digest which requires entropy from an RNG.

RandomizedMultipartSigner

Equivalent of RandomizedSigner but the message is provided in non-contiguous byte slices.

RandomizedMultipartSignerMut

Equivalent of RandomizedSignerMut but the message is provided in non-contiguous byte slices.

RandomizedSigner

Sign the given message using the provided external randomness source.

RandomizedSignerMut

Sign the provided message bytestring using &mut Self (e.g. an evolving cryptographic key such as a stateful hash-based signature), and a per-signature randomizer, returning a digital signature.

Signer

Sign the provided message bytestring using Self (e.g. a cryptographic key or connection to an HSM), returning a digital signature.

SignerMut

Sign the provided message bytestring using &mut Self (e.g. an evolving cryptographic key such as a stateful hash-based signature), returning a digital signature.