signature

Trait AsyncDigestSigner

pub trait AsyncDigestSigner<D, S>
where
    D: Update,
{
    // Required method
    async fn sign_digest_async<F: AsyncFn(&mut D) -> Result<(), Error>>(
        &self,
        f: F,
    ) -> Result<S, Error>;
}

Asynchronously sign the given prehashed message Digest using Self.

This trait is an async equivalent of the DigestSigner trait.

Required Methods

async fn sign_digest_async<F: AsyncFn(&mut D) -> Result<(), Error>>( &self, f: F, ) -> Result<S, Error>

Attempt to sign a message by updating the received Digest with it, returning a digital signature on success, or an error if something went wrong.

The given function can be invoked multiple times. It is expected that in each invocation the Digest is updated with the entire equal message.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.