signature

Trait AsyncVerifier

pub trait AsyncVerifier<S> {
    // Required method
    async fn verify_async(&self, msg: &[u8], signature: &S) -> Result<(), Error>;
}

Asynchronously verify the provided message bytestring using Self.

This trait is an async equivalent of the Verifier trait.

Required Methods

async fn verify_async(&self, msg: &[u8], signature: &S) -> Result<(), Error>

Asynchronously verify that the provided signature for a given message bytestring is authentic.

Returns Error if it is inauthentic, or otherwise returns ().

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.