Trait signature::Keypair

pub trait Keypair {
    type VerifyingKey: Clone;
    fn verifying_key(&self) -> Self::VerifyingKey;
}

Signing keypair with an associated verifying key.

This represents a type which holds both a signing key and a verifying key.

Associated Types

type VerifyingKey: Clone

Verifying key type for this keypair.

Required methods

fn verifying_key(&self) -> Self::VerifyingKey

Get the verifying key which can verify signatures produced by the signing key portion of this keypair.