Trait signature::RandomizedDigestSigner

pub trait RandomizedDigestSigner<D, S> where
    D: Digest,
    S: Signature, {
    fn try_sign_digest_with_rng(
        &self, 
        rng: impl CryptoRng + RngCore, 
        digest: D
    ) -> Result<S, Error>;

    fn sign_digest_with_rng(
        &self, 
        rng: impl CryptoRng + RngCore, 
        digest: D
    ) -> S { ... }
}

Combination of DigestSigner and RandomizedSigner with support for computing a signature over a digest which requires entropy from an RNG.

Required methods

fn try_sign_digest_with_rng(
&self,
rng: impl CryptoRng + RngCore,
digest: D
) -> Result<S, Error>

Attempt to sign the given prehashed message Digest, returning a digital signature on success, or an error if something went wrong.

Provided methods

fn sign_digest_with_rng(&self, rng: impl CryptoRng + RngCore, digest: D) -> S

Sign the given prehashed message Digest, returning a signature.

Panics in the event of a signing error.