signature
Module hazmat

Hazardous Materials: low-level APIs which can be insecure if misused.

The traits in this module are not generally recommended, and should only be used in special cases where they are specifically needed.

Using them incorrectly can introduce security vulnerabilities. Please carefully read the documentation before attempting to use them.

Traits

PrehashSigner

Sign the provided message prehash, returning a digital signature.

PrehashVerifier

Verify the provided message prehash using Self (e.g. a public key)