signature

Module signer

Traits for generating digital signatures

Traits

DigestSigner

Sign the given prehashed message Digest using Self.

RandomizedDigestSigner

Combination of DigestSigner and RandomizedSigner with support for computing a signature over a digest which requires entropy from an RNG.

RandomizedSigner

Sign the given message using the provided external randomness source.

Signer

Sign the provided message bytestring using Self (e.g. a cryptographic key or connection to an HSM), returning a digital signature.

SignerMut

Sign the provided message bytestring using &mut Self (e.g. an evolving cryptographic key such as a stateful hash-based signature), returning a digital signature.