Trait signature::DigestVerifier

pub trait DigestVerifier<D: Digest, S> {
    // Required method
    fn verify_digest(&self, digest: D, signature: &S) -> Result<(), Error>;
}

Verify the provided signature for the given prehashed message Digest is authentic.

Notes

This trait is primarily intended for signature algorithms based on the Fiat-Shamir heuristic, a method for converting an interactive challenge/response-based proof-of-knowledge protocol into an offline digital signature through the use of a random oracle, i.e. a digest function.

The security of such protocols critically rests upon the inability of an attacker to solve for the output of the random oracle, as generally otherwise such signature algorithms are a system of linear equations and therefore doing so would allow the attacker to trivially forge signatures.

To prevent misuse which would potentially allow this to be possible, this API accepts a Digest instance, rather than a raw digest value.

Required Methods

fn verify_digest(&self, digest: D, signature: &S) -> Result<(), Error>

Verify the signature against the given Digest output.